+++ src/actors/mod.rs
@@ -1,5 +1,4 @@
 pub mod sender_can;
 pub mod receiver_can;
 pub mod stdin;
-pub mod commands;
 pub mod ctrlc;
